requirements.txt
langchain
langchain-openai
langchain-community
langgraph
langsmith
faiss-cpu
pymupdf
fastapi
uvicorn[standard]
streamlit
pydantic-settings
python-multipart
plotly
pandas
httpx
sentence-transformers>=3.0.0
structlog
requests
python-dotenv
pyyaml
